clients/upsrw.c: fix receiver of strtol()

+++ clients/upsrw.c
@@ -317,7 +317,7 @@ static const char *get_data(const char *type, const char *varname)
 	return answer[3];
 }
 
-static void do_string(const char *varname, const int len)
+static void do_string(const char *varname, const long len)
 {
 	const char	*val;
 
@@ -328,7 +328,7 @@ static void do_string(const char *varname, const int len)
 	}
 
 	printf("Type: STRING\n");
-	printf("Maximum length: %d\n", len);
+	printf("Maximum length: %ld\n", len);
 	printf("Value: %s\n", val);
 }
 
@@ -352,7 +352,7 @@ static void do_number(const char *varname)
  * @param vartype the type of the NUT variable (ST_FLAG_STRING, ST_FLAG_NUMBER
  * @param len the length of the NUT variable, if type == ST_FLAG_STRING
  */
-static void do_enum(const char *varname, const int vartype, const int len)
+static void do_enum(const char *varname, const int vartype, const long len)
 {
 	int	ret;
 	unsigned int	numq, numa;
@@ -388,7 +388,7 @@ static void do_enum(const char *varname, const int vartype, const int len)
 		printf("Type: ENUM\n");
 
 	if (vartype == ST_FLAG_STRING)
-		printf("Maximum length: %d\n", len);
+		printf("Maximum length: %ld\n", len);
 
 	while (ret == 1) {
 
@@ -504,7 +504,7 @@ static void do_type(const char *varname)
 		if (!strncasecmp(answer[i], "STRING:", 7)) {
 
 			char	*len = answer[i] + 7;
-			int	length = strtol(len, NULL, 10);
+			long	length = strtol(len, NULL, 10);
 
 			if (is_enum == 1)
 				do_enum(varname, ST_FLAG_STRING, length);
